Move embed/qdrant/reduce/cluster clients under search/ with widened payload
- search/embed: embed.Client (llama.cpp /v1/embeddings)
- search/qdrant: QdrantClient, QdrantPoint (payload map[string]any), ScrolledPoint,
  Result; CollectionName → ProjectHash (returns bare 8-char hash)
- search/reduce: ReduceClient (PaCMAP/LocalMAP sidecar)
- search/cluster: HdbscanClient, Cluster, BuildClusters (depends on search/qdrant)
- internal/cli/search.go: updated imports; collection name built as
  "git-search-" + qdrant.ProjectHash(root) preserving existing names
- internal/search/{embed,qdrant,cluster}.go: removed
